src/tools_src/binarization/pyproject.toml
[tool.poetry]
name = "picturas-binarization-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that applies binarization filter"
authors = ["Marta Gonçalves <martagoncalves.inf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"
boto3 = "^1.35.98"
numpy = "^2.2.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
